backend/requirements.txt
Django==5.2.4
djangorestframework==3.15.2
django-cors-headers==4.3.1
drf-spectacular==0.27.2
Pillow==10.4.0
gunicorn==21.2.0
whitenoise==6.9.0
psycopg2-binary==2.9.9
dj-database-url==3.0.1
requests==2.31.0